Convert `long double` to `double` for cuda and hip etaMax

+++ inc/SYCLMath/GenVector/etaMax.h
@@ -45,8 +45,8 @@ namespace ROOT {
 #else
 __roohost__ __roodevice__ 
      long double etaMax_impl() {
-      return mylog ( std::numeric_limits<long double>::max()/256.0 ) -
-             mylog ( std::numeric_limits<long double>::denorm_min()*256.0 )
+      return mylog ( std::numeric_limits<double>::max()/256.0 ) -
+             mylog ( std::numeric_limits<double>::denorm_min()*256.0 )
              + 16.0 * log(2.0);
     // Actual usage of etaMax() simply returns the number 22756, which is
     // the answer this would supply, rounded to a higher integer.
